third_party/blink/web_tests/fast/scroll-snap/snaps-after-scrollbar-scrolling-thumb.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-scroll-snap-1" />
<link rel="stylesheet" href="resources/simple-snap.css">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>
<script src="../../resources/scrollbar-util.js"></script>
<script src="../../resources/blink-coordinates-util.js"></script>

<div id='scroller'>
  <div id="space"></div>
  <div class="snap left top"></div>
  <div class="snap right top"></div>
  <div class="snap left bottom"></div>
</div>

<script>
  if (window.internals) {
    internals.settings.setScrollAnimatorEnabled(false);
  }

  window.onload = async () => {
    // Skip test if there is no visible scrollbar thumb to drag.
    if (shouldSkipTests(
            !internals || internals.overlayScrollbarsEnabled,
            'Scrollbar is not visible')) {
      return;
    }

    await waitForCompositorCommit();

    const scroller = document.getElementById("scroller");
    const snapPoint = 400;
    const scrollDelta = 100;

    promise_test (async () => {
      await waitForScrollReset(scroller);
      const thumb = horizontalThumb(scroller);
      const scrollend = waitForScrollendEvent(scroller);
      await mouseDragAndDrop(thumb.x, thumb.y, thumb.x + scrollDelta, thumb.y);
      await scrollend;
      assert_approx_equals(scroller.scrollLeft, snapPoint, 1);
    }, "Snaps after dragging the horizontal scrollbar.");

    promise_test (async () => {
      await waitForScrollReset(scroller);
      const thumb = verticalThumb(scroller);
      const scrollend = waitForScrollendEvent(scroller);
      await mouseDragAndDrop(thumb.x, thumb.y, thumb.x, thumb.y + scrollDelta);
      await scrollend;
      assert_approx_equals(scroller.scrollTop, snapPoint, 1);
    }, "Snaps after dragging the vertical scrollbar.");
  };
</script>
